+++ core/src/main/kotlin/net/corda/core/transactions/TransactionBuilder.kt
@@ -45,7 +45,7 @@ import kotlin.collections.ArrayList
  * [TransactionState] with this notary specified will be generated automatically.
  */
 @DeleteForDJVM
-open class TransactionBuilder(
+open class TransactionBuilder @JvmOverloads constructor(
         var notary: Party? = null,
         var lockId: UUID = (Strand.currentStrand() as? FlowStateMachine<*>)?.id?.uuid ?: UUID.randomUUID(),
         protected val inputs: MutableList<StateRef> = arrayListOf(),
@@ -53,11 +53,11 @@ open class TransactionBuilder(
         protected val outputs: MutableList<TransactionState<ContractState>> = arrayListOf(),
         protected val commands: MutableList<Command<*>> = arrayListOf(),
         protected var window: TimeWindow? = null,
-        protected var privacySalt: PrivacySalt = PrivacySalt()
+        protected var privacySalt: PrivacySalt = PrivacySalt(),
+        protected val references: MutableList<StateRef> = arrayListOf()
 ) {
-    constructor(notary: Party) : this(notary, (Strand.currentStrand() as? FlowStateMachine<*>)?.id?.uuid ?: UUID.randomUUID())
-
     private val inputsWithTransactionState = arrayListOf<TransactionState<ContractState>>()
+    private val referencesWithTransactionState = arrayListOf<TransactionState<ContractState>>()
 
     /**
      * Creates a copy of the builder.
@@ -70,9 +70,11 @@ open class TransactionBuilder(
                 outputs = ArrayList(outputs),
                 commands = ArrayList(commands),
                 window = window,
-                privacySalt = privacySalt
+                privacySalt = privacySalt,
+                references = references
         )
         t.inputsWithTransactionState.addAll(this.inputsWithTransactionState)
+        t.referencesWithTransactionState.addAll(this.referencesWithTransactionState)
         return t
     }
 
@@ -82,6 +84,7 @@ open class TransactionBuilder(
         for (t in items) {
             when (t) {
                 is StateAndRef<*> -> addInputState(t)
+                is ReferencedStateAndRef<*> -> @Suppress("DEPRECATION") addReferenceState(t) // Will remove when feature finalised.
                 is SecureHash -> addAttachment(t)
                 is TransactionState<*> -> addOutputState(t)
                 is StateAndContract -> addOutputState(t.state, t.contract)
@@ -123,7 +126,7 @@ open class TransactionBuilder(
         }
 
         return SerializationFactory.defaultFactory.withCurrentContext(serializationContext) {
-            WireTransaction(WireTransaction.createComponentGroups(inputStates(), resolvedOutputs, commands, attachments + makeContractAttachments(services.cordappProvider), notary, window), privacySalt)
+            WireTransaction(WireTransaction.createComponentGroups(inputStates(), resolvedOutputs, commands, attachments + makeContractAttachments(services.cordappProvider), notary, window, referenceStates()), privacySalt)
         }
     }
 
@@ -137,6 +140,7 @@ open class TransactionBuilder(
      * TODO - review this logic
      */
     private fun makeContractAttachments(cordappProvider: CordappProvider): List<AttachmentId> {
+        // Reference inputs not included as it is not necessary to verify them.
         return (inputsWithTransactionState + outputs).map { state ->
             cordappProvider.getContractAttachmentID(state.contract)
                     ?: throw MissingContractAttachments(listOf(state))
@@ -155,42 +159,112 @@ open class TransactionBuilder(
         toLedgerTransaction(services).verify()
     }
 
-    open fun addInputState(stateAndRef: StateAndRef<*>): TransactionBuilder {
+    private fun checkNotary(stateAndRef: StateAndRef<*>) {
         val notary = stateAndRef.state.notary
-        require(notary == this.notary) { "Input state requires notary \"$notary\" which does not match the transaction notary \"${this.notary}\"." }
+        require(notary == this.notary) {
+            "Input state requires notary \"$notary\" which does not match the transaction notary \"${this.notary}\"."
+        }
+    }
+
+    // This check is performed here as well as in BaseTransaction.
+    private fun checkForInputsAndReferencesOverlap() {
+        val intersection = inputs intersect references
+        require(intersection.isEmpty()) {
+            "A StateRef cannot be both an input and a reference input in the same transaction."
+        }
+    }
+
+    private fun checkReferencesUseSameNotary() = referencesWithTransactionState.map { it.notary }.toSet().size == 1
+
+    /**
+     * Adds a reference input [StateRef] to the transaction.
+     *
+     * This feature was added in version 4 of Corda, so will throw an exception for any Corda networks with a minimum
+     * platform version less than 4.
+     *
+     * @throws UncheckedVersionException
+     */
+    @Deprecated(message = "Feature not yet released. Pending stabilisation.")
+    open fun addReferenceState(referencedStateAndRef: ReferencedStateAndRef<*>): TransactionBuilder {
+        val stateAndRef = referencedStateAndRef.stateAndRef
+        referencesWithTransactionState.add(stateAndRef.state)
+
+        // It is likely the case that users of reference states do not have permission to change the notary assigned
+        // to a reference state. Even if users _did_ have this permission the result would likely be a bunch of
+        // notary change races. As such, if a reference state is added to a transaction which is assigned to a
+        // different notary to the input and output states then all those inputs and outputs must be moved to the
+        // notary which the reference state uses.
+        //
+        // If two or more reference states assigned to different notaries are added to a transaction then it follows
+        // that this transaction likely _cannot_ be committed to the ledger as it unlikely that the party using the
+        // reference state can change the assigned notary for one of the reference states.
+        //
+        // As such, if reference states assigned to multiple different notaries are added to a transaction builder
+        // then the check below will fail.
+        check(checkReferencesUseSameNotary()) {
+            "Transactions with reference states using multiple different notaries are currently unsupported."
+        }
+
+        checkNotary(stateAndRef)
+        references.add(stateAndRef.ref)
+        checkForInputsAndReferencesOverlap()
+        return this
+    }
+
+    /** Adds an input [StateRef] to the transaction. */
+    open fun addInputState(stateAndRef: StateAndRef<*>): TransactionBuilder {
+        checkNotary(stateAndRef)
         inputs.add(stateAndRef.ref)
         inputsWithTransactionState.add(stateAndRef.state)
         return this
     }
 
+    /** Adds an attachment with the specified hash to the TransactionBuilder. */
     fun addAttachment(attachmentId: SecureHash): TransactionBuilder {
         attachments.add(attachmentId)
         return this
     }
 
+    /** Adds an output state to the transaction. */
     fun addOutputState(state: TransactionState<*>): TransactionBuilder {
         outputs.add(state)
         return this
     }
 
+    /** Adds an output state, with associated contract code (and constraints), and notary, to the transaction. */
     @JvmOverloads
-    fun addOutputState(state: ContractState, contract: ContractClassName, notary: Party, encumbrance: Int? = null, constraint: AttachmentConstraint = AutomaticHashConstraint): TransactionBuilder {
+    fun addOutputState(
+            state: ContractState,
+            contract: ContractClassName,
+            notary: Party, encumbrance: Int? = null,
+            constraint: AttachmentConstraint = AutomaticHashConstraint
+    ): TransactionBuilder {
         return addOutputState(TransactionState(state, contract, notary, encumbrance, constraint))
     }
 
     /** A default notary must be specified during builder construction to use this method */
     @JvmOverloads
-    fun addOutputState(state: ContractState, contract: ContractClassName, constraint: AttachmentConstraint = AutomaticHashConstraint): TransactionBuilder {
-        checkNotNull(notary) { "Need to specify a notary for the state, or set a default one on TransactionBuilder initialisation" }
+    fun addOutputState(
+            state: ContractState, contract: ContractClassName,
+            constraint: AttachmentConstraint = AutomaticHashConstraint
+    ): TransactionBuilder {
+        checkNotNull(notary) {
+            "Need to specify a notary for the state, or set a default one on TransactionBuilder initialisation"
+        }
         addOutputState(state, contract, notary!!, constraint = constraint)
         return this
     }
 
+    /** Adds a [Command] to the transaction. */
     fun addCommand(arg: Command<*>): TransactionBuilder {
         commands.add(arg)
         return this
     }
 
+    /**
+     * Adds a [Command] to the transaction, specified by the encapsulated [CommandData] object and required list of
+     * signing [PublicKey]s.
+     */
     fun addCommand(data: CommandData, vararg keys: PublicKey) = addCommand(Command(data, listOf(*keys)))
     fun addCommand(data: CommandData, keys: List<PublicKey>) = addCommand(Command(data, keys))
 
@@ -219,11 +293,19 @@ open class TransactionBuilder(
         return this
     }
 
-    // Accessors that yield immutable snapshots.
+    /** Returns an immutable list of input [StateRefs]. */
     fun inputStates(): List<StateRef> = ArrayList(inputs)
 
+    /** Returns an immutable list of reference input [StateRefs]. */
+    fun referenceStates(): List<StateRef> = ArrayList(references)
+
+    /** Returns an immutable list of attachment hashes. */
     fun attachments(): List<SecureHash> = ArrayList(attachments)
+
+    /** Returns an immutable list of output [TransactionState]s. */
     fun outputStates(): List<TransactionState<*>> = ArrayList(outputs)
+
+    /** Returns an immutable list of [Command]s. */
     fun commands(): List<Command<*>> = ArrayList(commands)
 
     /**