+++ node/src/main/kotlin/net/corda/node/services/transactions/RaftUniquenessProvider.kt
@@ -200,14 +200,17 @@ class RaftUniquenessProvider(
             txId: SecureHash,
             callerIdentity: Party,
             requestSignature: NotarisationRequestSignature,
-            timeWindow: TimeWindow?) {
+            timeWindow: TimeWindow?,
+            references: List<StateRef>
+    ) {
         log.debug { "Attempting to commit input states: ${states.joinToString()}" }
         val commitCommand = CommitTransaction(
                 states,
                 txId,
                 callerIdentity.name.toString(),
                 requestSignature.serialize().bytes,
-                timeWindow
+                timeWindow,
+                references
         )
         val commitError = client.submit(commitCommand).get()
         if (commitError != null) throw NotaryInternalException(commitError)